update packages and init TeamPresenter.kt
