+++ app/src/main/java/com/example/busybee/ui/home/HomeViewPagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.busybee.ui.home.view
+package com.example.busybee.ui.home
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
