+++ app/src/main/java/com/example/busybee/ui/home/personalTask/doneTask/DoneAdapter.kt
@@ -1,26 +1,26 @@
-package com.example.busybee.ui.home.view.teamTask.view.inProgress
+package com.example.busybee.ui.home.personalTask.doneTask
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.example.busybee.base.BaseAdapter
-import com.example.busybee.data.models.TeamToDo
+import com.example.busybee.data.models.PersonalTodo
 import com.example.busybee.databinding.ItemTaskBinding
 
-class TeamInProgressAdapter(private var teamInProgressList: List<TeamToDo>) :
-    BaseAdapter<TeamToDo, ItemTaskBinding>(teamInProgressList) {
+class DoneAdapter(personalToDoList: List<PersonalTodo>) :
+    BaseAdapter<PersonalTodo, ItemTaskBinding>(personalToDoList) {
+
     override val bindingInflater: (LayoutInflater, ViewGroup?, Boolean) -> ItemTaskBinding
         get() = ItemTaskBinding::inflate
 
     override fun onBindViewHolder(
         holder: BaseViewHolder<ItemTaskBinding>,
         position: Int,
-        currentItem: TeamToDo,
+        currentItem: PersonalTodo
     ) {
         with(holder.binding) {
             textTaskName.text = currentItem.title
             textTaskDescription.text = currentItem.description
             textTaskTime.text = currentItem.creationTime
         }
-
     }
 }
