+++ app/src/main/java/com/example/busybee/ui/home/personalTask/inProgressTask/PersonalInProgressFragment.kt
@@ -1,8 +1,8 @@
-package com.example.busybee.ui.home.view.personalTask.view.inProgressTask
+package com.example.busybee.ui.home.personalTask.inProgressTask
 
 import com.example.busybee.base.BaseFragment
 import com.example.busybee.databinding.FragmentPersonalInProgressBinding
-import com.example.busybee.ui.home.view.personalTask.view.inProgressTask.PersonalInProgressAdapter
+import com.example.busybee.ui.home.personalTask.inProgressTask.PersonalInProgressAdapter
 
 class PersonalInProgressFragment : BaseFragment<FragmentPersonalInProgressBinding>() {
     private lateinit var adapter : PersonalInProgressAdapter
