+++ app/src/main/java/com/example/busybee/ui/home/personalTask/toDoTask/PersonalToDoAdapter.kt
@@ -1,28 +1,26 @@
-package com.example.busybee.ui.home.view.teamTask.view.toDo
+package com.example.busybee.ui.home.personalTask.toDoTask
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.example.busybee.base.BaseAdapter
-import com.example.busybee.data.models.TeamToDo
+import com.example.busybee.data.models.PersonalTodo
 import com.example.busybee.databinding.ItemTaskBinding
 
-class TeamToDoAdapter(private var teamToDoList: List<TeamToDo>) :
-    BaseAdapter<TeamToDo, ItemTaskBinding>(teamToDoList) {
+class PersonalToDoAdapter(personalToDoList: List<PersonalTodo>) :
+    BaseAdapter<PersonalTodo, ItemTaskBinding>(personalToDoList) {
+
     override val bindingInflater: (LayoutInflater, ViewGroup?, Boolean) -> ItemTaskBinding
         get() = ItemTaskBinding::inflate
 
     override fun onBindViewHolder(
         holder: BaseViewHolder<ItemTaskBinding>,
         position: Int,
-        currentItem: TeamToDo,
+        currentItem: PersonalTodo
     ) {
         with(holder.binding) {
             textTaskName.text = currentItem.title
             textTaskDescription.text = currentItem.description
             textTaskTime.text = currentItem.creationTime
         }
-
-
+        }
     }
-
-}
