--- app/src/main/java/com/example/busybee/ui/home/presenter/HomePresenter.kt
@@ -1,5 +1,0 @@
-package com.example.busybee.ui.home.presenter
-
-
-class HomePresenter {
-}
