+++ app/src/main/java/com/example/busybee/ui/home/teamtask/presenter/TeamPresenter.kt
@@ -1,0 +1,6 @@
+package com.example.busybee.ui.home.teamtask.presenter
+
+import com.example.busybee.data.RepositoryInterface
+
+class TeamPresenter(private val repository: RepositoryInterface) {
+}
