+++ app/src/main/java/com/example/busybee/ui/home/teamtask/presenter/TeamPresenterInterface.kt
@@ -1,0 +1,4 @@
+package com.example.busybee.ui.home.teamtask.presenter
+
+interface TeamPresenterInterface {
+}
