+++ app/src/main/java/com/example/busybee/ui/home/teamtask/view/done/TeamDoneFragment.kt
@@ -1,4 +1,4 @@
-package com.example.busybee.ui.home.view.teamTask.view.done
+package com.example.busybee.ui.home.teamtask.view.done
 
 import com.example.busybee.base.BaseFragment
 import com.example.busybee.databinding.FragmentTeamDoneBinding
