+++ app/src/main/java/com/example/busybee/ui/home/teamtask/view/inProgress/TeamInProgressAdapter.kt
@@ -1,26 +1,26 @@
-package com.example.busybee.ui.home.view.personalTask.view.doneTask
+package com.example.busybee.ui.home.teamtask.view.inProgress
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.example.busybee.base.BaseAdapter
-import com.example.busybee.data.models.PersonalTodo
+import com.example.busybee.data.models.TeamToDo
 import com.example.busybee.databinding.ItemTaskBinding
 
-class DoneAdapter(personalToDoList: List<PersonalTodo>) :
-    BaseAdapter<PersonalTodo, ItemTaskBinding>(personalToDoList) {
-
+class TeamInProgressAdapter(private var teamInProgressList: List<TeamToDo>) :
+    BaseAdapter<TeamToDo, ItemTaskBinding>(teamInProgressList) {
     override val bindingInflater: (LayoutInflater, ViewGroup?, Boolean) -> ItemTaskBinding
         get() = ItemTaskBinding::inflate
 
     override fun onBindViewHolder(
         holder: BaseViewHolder<ItemTaskBinding>,
         position: Int,
-        currentItem: PersonalTodo
+        currentItem: TeamToDo,
     ) {
         with(holder.binding) {
             textTaskName.text = currentItem.title
             textTaskDescription.text = currentItem.description
             textTaskTime.text = currentItem.creationTime
         }
+
     }
 }
