+++ app/src/main/res/layout/fragment_personal_in_progress.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.view.personalTask.view.inProgressTask.PersonalInProgressFragment">
+    tools:context=".ui.home.personalTask.inProgressTask.PersonalInProgressFragment">
 
     <include
         android:id="@+id/header_inProgress"
