+++ app/src/main/res/layout/fragment_personal_to_do.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/white_300"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.home.view.personalTask.view.toDoTask.PersonalToDoFragment">
+    tools:context=".ui.home.personalTask.toDoTask.PersonalToDoFragment">
 
     <include
         android:id="@+id/header_ToDo"
